requirements.txt
tensorflow==2.18.0
numpy
pandas
tqdm
python-chess